requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
astropy==6.1.0
astropy-iers-data==0.2024.6.10.0.30.47
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
decorator==5.1.1
decouple==0.0.7
Django==5.1.1
django-bootstrap3==23.4
django-braces==1.15.0
django-debug-toolbar==4.2.0
django-ratelimit==4.1.0
ephem==4.1.5
Faker==19.11.0
Flask==2.3.3
frozenlist==1.4.0
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.0
h11==0.14.0
h3==3.7.7
idna==3.4
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
misaka==2.1.1
multidict==6.0.4
numpy==1.26.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.1.3
Pillow==10.1.0
protobuf==4.25.0
pycparser==2.21
pycryptodomex==3.19.0
pyerfa==2.0.1.4
pyseto==1.7.5
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
ratelim==0.1.6
replit==3.3.2
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.1
timezonefinder==6.5.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2024.2
urllib3==1.26.15
Werkzeug==2.3.8
wsproto==1.2.0
yarl==1.9.2
gunicorn==23.0.0
python-decouple==3.8
django-ckeditor-5==0.2.13
django-cors-headers==4.6.0
djangorestframework==3.15.2
django-rest-framework==0.1.0